+++ internal/tui/tui.go
@@ -77,7 +77,7 @@ func Logo(version string) {
  ╚═════╝ ╚═╝  ╚═╝ ╚═════╝╚═╝  ╚═╝╚═╝     ╚═╝  ╚═╝ ╚═════╝╚═╝  ╚═╝`)
 	fmt.Print(Reset)
 	fmt.Printf("%s Arange-tun  %s%s%s\n", Bold+White, Red, version, Reset)
-	fmt.Println(Gray + " Telegram : @devmahdi_com  |  GitHub : https://github.com/mahdi-byte64" + Reset)
+	fmt.Println(Gray + " GitHub : https://github.com/mahdi-byte64" + Reset)
 }
 
 // Prompt reads a trimmed line after printing label.